refactor(cli): migrate fdsn + firms tooling onto discovery
Move fdsn (refresher/curated_ids/validator) and firms (refresher/prober/
validator, with the map-key redaction) to earthlens.<backend>.cli, register them
in the hazards earthlens.cli table, and drop their entries + functions from
core's refresh/curate/validate dicts. Add get_text/redact/infer_dtype/HTTP_TIMEOUT
to the public toolkit. Relocate their CLI tests to the hazards distribution.
Shrink the guard allow-list by two.
